move helper functions from SubgraphWrapper to graph/utils, move deciding whether to render from 'branch' to Results render method
